remove usage of separate environment in os cli than os processes
