+++ data-node/src/main/java/org/graylog/datanode/opensearch/cli/OpensearchCommandLineProcess.java
@@ -23,13 +23,11 @@ import com.github.rholder.retry.RetryerBuilder;
 import com.github.rholder.retry.StopStrategies;
 import com.github.rholder.retry.WaitStrategies;
 import jakarta.validation.constraints.NotNull;
-import org.apache.commons.exec.OS;
 import org.graylog.datanode.configuration.OpensearchConfigurationDir;
 import org.graylog.datanode.configuration.OpensearchConfigurationException;
 import org.graylog.datanode.opensearch.configuration.OpensearchConfiguration;
 import org.graylog.datanode.process.CommandLineProcess;
 import org.graylog.datanode.process.CommandLineProcessListener;
-import org.graylog.datanode.process.Environment;
 import org.graylog.datanode.process.ProcessInformation;
 import org.graylog.datanode.process.ProcessListener;
 import org.graylog.datanode.process.configuration.beans.OpensearchKeystoreItem;
@@ -37,13 +35,9 @@ import org.graylog.datanode.process.configuration.files.DatanodeConfigFile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.BufferedReader;
 import java.io.Closeable;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.nio.charset.Charset;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Collection;
 import java.util.List;
@@ -55,36 +49,6 @@ public class OpensearchCommandLineProcess implements Closeable {
 
     private final CommandLineProcess commandLineProcess;
     private final CommandLineProcessListener resultHandler;
-
-
-    /**
-     * as long as OpenSearch is not supported on macOS, we have to fix the jdk path if we want to
-     * start the DataNode inside IntelliJ.
-     */
-    private Environment fixJdkOnMac(Environment environment) {
-        if (!OS.isFamilyMac()) {
-            return environment;
-        }
-        // set system jdk as OS jdk
-        final ProcessBuilder builder = new ProcessBuilder("/usr/libexec/java_home");
-        builder.redirectErrorStream(true);
-        try {
-            final Process process = builder.start();
-            final BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream(), Charset.defaultCharset()));
-            var line = reader.readLine();
-            if (line != null && Files.exists(Path.of(line))) {
-                return environment.withOpensearchJavaHome(Path.of(line));
-            } else {
-                LOG.error("Output of '/usr/libexec/java_home' is not the jdk: {}", line);
-            }
-            // cleanup
-            process.destroy();
-            reader.close();
-        } catch (IOException e) {
-            LOG.error("Could not link jdk.app on macOS: {}", e.getMessage(), e);
-        }
-        return environment;
-    }
 
     private void writeOpenSearchConfig(final OpensearchConfiguration config) {
         final OpensearchConfigurationDir confDir = config.getOpensearchConfigurationDir();
@@ -103,13 +67,12 @@ public class OpensearchCommandLineProcess implements Closeable {
     }
 
     public OpensearchCommandLineProcess(OpensearchConfiguration config, ProcessListener listener) {
-        Environment env = fixJdkOnMac(config.getEnv());
         configureOpensearchKeystoreSecrets(config);
         final Path executable = config.getOpensearchDistribution().getOpensearchExecutable();
         writeOpenSearchConfig(config);
         logWarnings(config);
         resultHandler = new CommandLineProcessListener(listener);
-        commandLineProcess = new CommandLineProcess(executable, List.of(), resultHandler, env);
+        commandLineProcess = new CommandLineProcess(executable, List.of(), resultHandler, config.getEnv());
     }
 
     private void logWarnings(OpensearchConfiguration config) {
