+++ data-node/src/test/java/org/graylog/datanode/opensearch/cli/OpensearchKeystoreCommandLineIT.java
@@ -21,6 +21,7 @@ import org.assertj.core.api.Assertions;
 import org.graylog.datanode.OpensearchDistribution;
 import org.graylog.datanode.configuration.OpensearchArchitecture;
 import org.graylog.datanode.configuration.OpensearchDistributionProvider;
+import org.graylog.datanode.process.Environment;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
@@ -49,7 +50,10 @@ class OpensearchKeystoreCommandLineIT {
 
     private OpensearchCli createCli(Path tempDir) throws URISyntaxException {
         final Path binDirPath = detectOpensearchBinDir();
-        return new OpensearchCli(new CliEnv(tempDir), binDirPath);
+        Environment env = new Environment(System.getenv())
+                .withOpensearchJavaHome(Path.of(System.getProperty("java.home")))
+                .withOpensearchPathConf(tempDir);
+        return new OpensearchCli(env, binDirPath);
     }
 
     @Nonnull
